% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{accounts}
\alias{accounts}
\title{accounts}
\description{
class for account management
}
\details{
Methods for creating a user accounts dataset for creating basic
"password protected" Shiny application. These methods aren't designed
for any rebust applications and should be used for offering very basic
authentication. It is not recommended to use these methods for use in
applications that store any senstitive information.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-config}{\code{accounts$config()}}
\item \href{#method-add_user}{\code{accounts$add_user()}}
\item \href{#method-remove_user}{\code{accounts$remove_user()}}
\item \href{#method-reset_password}{\code{accounts$reset_password()}}
\item \href{#method-print_data}{\code{accounts$print_data()}}
\item \href{#method-save_data}{\code{accounts$save_data()}}
\item \href{#method-clone}{\code{accounts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-config"></a>}}
\if{latex}{\out{\hypertarget{method-config}{}}}
\subsection{Method \code{config()}}{
config
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$config(name, version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the accounts database}

\item{\code{version}}{db version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_user"></a>}}
\if{latex}{\out{\hypertarget{method-add_user}{}}}
\subsection{Method \code{add_user()}}{
add_user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$add_user(username, password, type = "standard")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{account name}

\item{\code{password}}{account password}

\item{\code{type}}{account type; developer defined}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_user"></a>}}
\if{latex}{\out{\hypertarget{method-remove_user}{}}}
\subsection{Method \code{remove_user()}}{
remove_user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$remove_user(username)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{name for the account}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_password"></a>}}
\if{latex}{\out{\hypertarget{method-reset_password}{}}}
\subsection{Method \code{reset_password()}}{
reset_password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$reset_password(username, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{account name}

\item{\code{password}}{account password}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_data"></a>}}
\if{latex}{\out{\hypertarget{method-print_data}{}}}
\subsection{Method \code{print_data()}}{
print dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$print_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_data"></a>}}
\if{latex}{\out{\hypertarget{method-save_data}{}}}
\subsection{Method \code{save_data()}}{
save data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$save_data(dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{output path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{accounts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
