% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{back_btn}
\alias{back_btn}
\title{Back Button}
\usage{
back_btn(inputId = "backBtn", label = "Previous")
}
\arguments{
\item{inputId}{a unique Id for the button (default \code{backBtn})}

\item{label}{customize the button (default \code{Previous})}
}
\description{
Create a back button for use in the navigation container. By default,
this button renders with the inputId defined as \code{backBtn}. The button can
be accessed in the shiny server via \code{input$backBtn}. You can also set the
button label. The default is \code{Previous}.
}
