% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{Card}
\usage{
card(
  inputId,
  text,
  icon_name = "check_circle",
  icon_type = "outline",
  class = NULL
)
}
\arguments{
\item{inputId}{a unique ID for the card}

\item{text}{text or HTML string to display}

\item{icon_name}{an rheroicons icon name}

\item{icon_type}{an rheroicons icon type ("output" or "solid")}

\item{class}{css classes to apply to the card}
}
\description{
Create a generic text card that displays an icon and title side by side.
The title can be changed from the shiny server using the function
\code{update_card}.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- tagList(
    iceComponents::use_iceComponents(),
    iceComponents::container(
      iceComponents::page(
        tags$h2("Study Progress", style = "text-align: center;"),
        iceComponents::card(
          inputId = "experiments",
          text = as.character(
            tagList(
              tags$strong("200"),
              "experiments performed"
            )
          ),
          icon_name = "beaker",
          icon_type = "solid"
        )
      )
    )
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}

}
\references{
\url{https://github.com/davidruvolo51/rheroicons}
}
