% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_group.R
\name{card_group}
\alias{card_group}
\title{Container for cards}
\usage{
card_group(..., class = NULL)
}
\arguments{
\item{...}{cards and other UI content}

\item{class}{a string containing one or more css classes}
}
\description{
A layout container for grouping more than one card
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- tagList(
    iceComponents::use_iceComponents(),
    iceComponents::container(
      iceComponents::page(
        tags$h2("Study Progress", style = "text-align: center;"),
        iceComponents::card_group(
          iceComponents::card(
            inputId = "experiments",
            text = as.character(
              tagList(
                tags$strong("200"),
                "experiments performed"
              )
            ),
            icon_name = "beaker",
            icon_type = "solid"
          ),
          iceComponents::card(
            inputId = "recruitment",
            text = as.character(
              tagList(
                "We recruited",
                tags$strong("62"),
                "new participants"
              )
            ),
            icon_name = "user_add",
            icon_type = "outline"
          ),
          iceComponents::card(
            inputId = "statisfaction",
            text = as.character(
              tagList(
                tags$strong("98.9\%"),
                "study satisfaction rate!"
              )
            ),
            icon_name = "emoji_happy",
            icon_type = "outline"
          )
        )
      )
    )
  )
}

}
