% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_group.R
\name{clear_checkbox_group}
\alias{clear_checkbox_group}
\title{Clear Checkbox Group}
\usage{
clear_checkbox_group(inputId)
}
\arguments{
\item{inputId}{ID of the Checkbox group to clear}
}
\description{
Clear all checked options
}
\keyword{checkboxgroup}
\keyword{clear}
\keyword{iceComponetns}
