% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fields.R
\name{clear_input}
\alias{clear_input}
\title{Clear Input}
\usage{
clear_input(inputId)
}
\arguments{
\item{inputId}{Id of the component to clear}
}
\description{
Remove all invalidation attributes from an input element while
leaving user entered input text untouched. Use \code{reset_input} for
removing everything.
}
