% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{container}
\alias{container}
\title{Container}
\usage{
container(..., classnames = NULL)
}
\arguments{
\item{...}{Shiny UI content}

\item{classnames}{optional css classes to apply to the container}
}
\description{
Functional component for creating a new app container. Use \code{page_ui},
to define individual pages.
}
\examples{
if (interactive()) {
  library(shiny)
  library(iceComponents)
  ui <- tagList(
    use_iceComponents(),
    container(
      classnames = "dark-theme",
      tags$h1("My Application")
    )
  )
  server <- function(input, output, session) {
  }
  shinyApp(ui, server)
}

}
