% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_box.R
\name{error_box}
\alias{error_box}
\title{error_box}
\usage{
error_box(
  inputId,
  class = NULL,
  icon_name = "exclamation",
  icon_type = "outline"
)
}
\arguments{
\item{inputId}{a unique id for the error box}

\item{class}{pass css class names}

\item{icon_name}{an rheroicons icon name}

\item{icon_type}{an rheroicons icon type ("output" or "solid")}
}
\description{
Create an error message box to inform users when an error has occured.
This element is designed to render without an message. Use the server-side
function \code{show_error_box} to render an message into the error box, and then
the \code{hide_error_box} to reset the error. You can further customize the error
box to fit the error by choosing an icon made available by the \code{rheroicons}
package.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- tagList(
    iceComponents::use_iceComponents(),
    iceComponents::container(
      iceComponents::page(
        tags$h2("My App"),
        tags$p("Please wait while the app loads..."),
        iceComponents::error_box(
          inputId = "shiny-server-error"
        )
      )
    )
  )

  server <- function() {
    err <- function() {
       Sys.sleep(2)
       iceComponents::show_error_box(
         inputId = "shiny-server-error",
         error = "Internal Server Error (500)"
       )
    }

    err()

  }

  shinyApp(ui, server)

}

}
\references{
\url{https://github.com/davidruvolo51/rheroicons}
}
