% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{forward_btn}
\alias{forward_btn}
\title{Forward Button}
\usage{
forward_btn(inputId = "forwardBtn", label = "Next")
}
\arguments{
\item{inputId}{a unique Id for the button (default \code{forwardBtn})}

\item{label}{set the button label (default \code{Next})}
}
\description{
Create a forward button for use in the navigation container. By default,
this button renders with the inputId defined as \code{forwardBtn}. The button can
be accessed in the shiny server via \code{input$forwardBtn}. You can also set the
button label. The default is \code{Next}.
}
