% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fields.R
\name{input}
\alias{input}
\title{Input}
\usage{
input(inputId, label, type = "text", icon = NULL)
}
\arguments{
\item{inputId}{unique ID for the component}

\item{label}{text that describes the input}

\item{type}{input type (password or text)}

\item{icon}{an rheroicons icon}
}
\description{
Create a input field for use in Shiny. This component returns an
input element (either password or text) with the addition of functions for
validating inputs
}
