% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{invalidate_input}
\alias{invalidate_input}
\title{Invalidate Input}
\usage{
invalidate_input(inputId, error)
}
\arguments{
\item{inputId}{Id of the password component to invalidate}

\item{error}{message to display that describes the error}
}
\description{
Invalidate an input field based on server side logic
}
