% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{navigation}
\alias{navigation}
\title{Page Navigation}
\usage{
navigation(...)
}
\arguments{
\item{...}{Navigation buttons (\code{back_btn}, \code{forward_btn})}
}
\description{
Create a container for page navigation buttons. The \code{navigation} function
returns a container for the page navigation buttons.
}
\examples{
if (interactive()) {
  library(shiny)
  library(iceComponents)

  # define app
  ui <- tagList(
    use_iceComponents(),
    container(
      tags$h2("Count by 5's"),
      uiOutput("count"),
      navigation(
        back_btn(inputId = "subtract", label = "Subtract 5"),
        back_btn(inputId = "add", label = "Add 5"),
      )
    )
  )

  # dynamically render pages based on page navigation
  server <- function(input, output) {

    counter <- reactiveVal(1)

    observeEvent(input$subtract, {
      counter(counter() - 5)
      output$counter <- renderUI(tags$p(counter()))
    })

    observeEvent(input$add, {
      counter(counter() + 5)
      output$count <- renderUI(tags$p(counter()))
    })

  }

  shinyApp(ui, server)
}

}
