% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{page}
\alias{page}
\title{Page Container}
\usage{
page(..., inputId, classnames = NULL)
}
\arguments{
\item{...}{Shiny UI content}

\item{inputId}{an ID for the page (required)}

\item{classnames}{apply css classes (optional)}
}
\description{
Define a "page" region inside the main container
}
\examples{
if (interactive()) {
  library(shiny)
  library(iceComponents)
  ui <- tagList(
    use_iceComponents(),
    container(
      classnames = "dark-theme",
      page(
        tags$h1("My Application")
      )
    )
  )
  server <- function(input, output, session) {
  }
  shinyApp(ui, server)
}

}
