% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_input.R
\name{reset_accordion_input}
\alias{reset_accordion_input}
\title{\code{reset_accordion_input}}
\usage{
reset_accordion_input(inputId)
}
\arguments{
\item{inputId}{the of of the component to reset}
}
\description{
A server-side function that resets the accordion input component to it's
default statue. Resetting the component will uncheck the import, close
the accordion (if open), and reset the input value.
}
