% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_group.R
\name{reset_checkbox_group}
\alias{reset_checkbox_group}
\title{Reset Checkbox Group}
\usage{
reset_checkbox_group(inputId)
}
\arguments{
\item{inputId}{ID of the Checkbox group to reset}
}
\description{
Reset a checkbox group to it's initial state.
}
\keyword{checkboxgroup}
\keyword{iceComponents}
\keyword{reset}
