% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_doc_attribs.R
\name{set_doc_attribs}
\alias{set_doc_attribs}
\title{Set Document Attributes}
\usage{
set_doc_attribs(lang = "en", dir = "ltr")
}
\arguments{
\item{lang}{the language the document is written in (ie., ISO639 code)}

\item{dir}{the direction of the content (e.g., left to right)}
}
\description{
Easily set document language and direction
}
\examples{
if (interactive()) {
  ui <- tagList(
    iceComponents::use_iceComponents(),
    iceComponents::set_doc_attribs(lang = "en", dir = "ltr"),
    iceComponents::main(
      tags$h1("This is my cool app")
    )
  )
}
}
