% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_box.R
\name{show_error_box}
\alias{show_error_box}
\title{show_error_box}
\usage{
show_error_box(inputId, error)
}
\arguments{
\item{inputId}{ID of error to update}

\item{error}{message to display}
}
\description{
Reveal a message box and display a message
}
