% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_text.R
\name{show_error_text}
\alias{show_error_text}
\title{show_error_text}
\usage{
show_error_text(inputId, error)
}
\arguments{
\item{inputId}{ID of the error element to display}

\item{error}{message to display}
}
\description{
Display an error text message
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- tagList(
    iceComponents::use_iceComponents(),
    iceComponents::container(
      iceComponents::page(
        tags$h2("My App"),
        iceComponents::error_text(
          inputId = "shiny-server-error"
        ),
        tags$button(
          id = "show",
          class = "shiny-bound-input action-button",
          "Do Something"
        )
      )
    )
  )

  server <- function() {
    observeEvent(input$show, {
      iceComponents::show_error_text(
        inputId = "shiny-server-error",
        error = "Internal Server Error (500)"
      )
    })
  }

  shinyApp(ui, server)

}


}
