% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{update_card}
\alias{update_card}
\title{Update Card}
\usage{
update_card(inputId, text)
}
\arguments{
\item{inputId}{the ID of the card to update}

\item{text}{a string containing a title to send to the card}
}
\description{
Update the text of card
}
