% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_iceComponents.R
\name{use_iceComponents}
\alias{use_iceComponents}
\title{use_iceComponents}
\usage{
use_iceComponents()
}
\description{
Client-side function for initiating package resources
}
\examples{
if(interactive()) {
  library(shiny)
  ui <- tagList(
    use_iceComponents(),
    tags$main(
      tags$h2("Hello, world!")
    )
  )
  server <- function(input, output) {}
  shinyApp(ui, server)
}
}
