% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_group.R
\name{checkbox_group}
\alias{checkbox_group}
\title{checkbox_group}
\usage{
checkbox_group(
  inputId,
  title,
  caption = NULL,
  choices,
  values = NULL,
  checked = FALSE
)
}
\arguments{
\item{inputId}{a unique id for the input group}

\item{title}{a title for the inputs}

\item{caption}{an optional caption to display}

\item{choices}{an array of choices}

\item{values}{an array of values to display}

\item{checked}{an array of logical values to check an input at start}
}
\description{
Create a custom checkbox input group
}
\keyword{checkboxgroup}
\keyword{iceComponents}
