% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_box.R
\name{show_error_box}
\alias{show_error_box}
\title{show_error_box}
\usage{
show_error_box(inputId, error)
}
\arguments{
\item{inputId}{ID of error to update}

\item{error}{message to display}
}
\description{
Reveal a message box and display a message
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- tagList(
    iceComponents::use_iceComponents(),
    iceComponents::container(
      iceComponents::page(
        tags$h2("My App"),
        iceComponents::error_box(
          inputId = "shiny-server-error"
        ),
        tags$button(
          id = "show",
          class = "shiny-bound-input action-button",
          "Do Something"
        )
      )
    )
  )

  server <- function() {
    observeEvent(input$show, {
      iceComponents::show_error_box(
        inputId = "shiny-server-error",
        error = "Internal Server Error (500)"
      )
    })
  }

  shinyApp(ui, server)

}

}
